+++ src/constants/index.ts
@@ -13,3 +13,4 @@ export * from './JULIES_BACKGROUND';
 export * from './JACKIES_BACKGROUND';
 export * from './DEV1_CONFIG';
 export * from './GITHUB_WEBHOOK';
+export * from './GITHUB_DEPLOY_MONITORS';