+++ src/types/github/index.ts
@@ -1,1 +1,2 @@
 export * from './GitHubWebhook';
+export * from './GitHubDeployStatus';